python/requirements.txt
boto3==1.9.131
botocore==1.12.131
unittest-xml-reporting==2.5.1
asn1crypto==0.24.0
azure-common==1.1.19
azure-storage-blob==1.5.0
azure-storage-common==1.4.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
future==0.17.1
idna==2.8
ijson==2.3
pyOpenSSL==19.0.0
pycparser==2.19
pycryptodomex==3.8.1
pyjwt==1.7.1
requests==2.21.0
snowflake-connector-python==1.9.1